requirements.txt
agora-token-builder==1.0.0
asgiref==3.7.2
attrs==23.1.0
autopep8==1.5.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
Django==4.2.6
django-cors-headers==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
MarkupSafe==2.1.3
ndg-httpsclient==0.5.1
packaging==23.2
Pillow==8.1.0
pusher==3.0.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.8.0
PyMySQL==1.1.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==6.0.1
referencing==0.31.0
requests==2.25.1
rpds-py==0.13.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
six==1.15.0
sqlparse==0.4.1
stripe==6.7.0
toml==0.10.2
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.3
whitenoise==6.5.0
